@@ -1,15 +1,17 @@
 {
   "name": "holper",
   "version": "0.0.1",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "preview": "vite preview --port 5050",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
     "test:unit": "vitest",
     "typecheck": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts,.mdx --fix --ignore-path .gitignore",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "storybook dev",
+    "build-storybook": "storybook build"
   },
   "dependencies": {
     "pinia": "^2.0.13",
@@ -17,37 +19,39 @@
     "vue-router": "^4.0.14"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-svg-core": "~6.2.0",
-    "@fortawesome/free-solid-svg-icons": "~6.2.0",
+    "@fortawesome/fontawesome-svg-core": "~6.5.0",
+    "@fortawesome/free-solid-svg-icons": "~6.5.0",
     "@fortawesome/vue-fontawesome": "^3.0.1",
-    "@babel/core": "^7.19.1",
     "@rushstack/eslint-patch": "^1.1.0",
-    "@storybook/addon-actions": "^6.5.12",
-    "@storybook/addon-essentials": "^6.5.12",
-    "@storybook/addon-interactions": "^6.5.12",
-    "@storybook/addon-links": "^6.5.12",
-    "@storybook/builder-vite": "^0.2.2",
-    "@storybook/testing-library": "^0.0.13",
-    "@storybook/vue3": "^6.5.12",
-    "@types/node": "^18.7.18",
-    "@vitejs/plugin-vue": "^3.1.0",
-    "@vue/eslint-config-prettier": "^7.0.0",
-    "@vue/eslint-config-typescript": "^11.0.1",
-    "@vue/test-utils": "^2.0.0-rc.18",
-    "@vue/tsconfig": "^0.1.3",
-    "babel-loader": "^8.2.5",
-    "chromatic": "^6.9.0",
+    "@storybook/addon-actions": "^7.6.6",
+    "@storybook/addon-essentials": "^7.6.6",
+    "@storybook/addon-interactions": "^7.6.6",
+    "@storybook/addon-links": "^7.6.6",
+    "@storybook/testing-library": "^0.2.2",
+    "@storybook/vue3": "^7.6.6",
+    "@storybook/vue3-vite": "^7.6.6",
+    "@tsconfig/node18": "^18.2.2",
+    "@types/node": "^20.0.0",
+    "@vitejs/plugin-vue": "^5.0.0",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/test-utils": "^2.4.3",
+    "@vue/tsconfig": "^0.5.1",
+    "autoprefixer": "^10.4.16",
+    "chromatic": "^10.2.0",
     "eslint": "^8.5.0",
-    "eslint-plugin-mdx": "^2.0.5",
-    "eslint-plugin-storybook": "^0.6.4",
+    "eslint-plugin-mdx": "^3.1.1",
+    "eslint-plugin-storybook": "^0.6.15",
     "eslint-plugin-vue": "^9.5.1",
-    "happy-dom": "^6.0.4",
-    "prettier": "^2.5.1",
+    "happy-dom": "^12.10.3",
+    "prettier": "^3.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "storybook": "^7.6.6",
     "tailwindcss": "^3.1.8",
-    "typescript": "~4.8.3",
-    "vite": "^3.1.3",
-    "vitest": "^0.8.1",
-    "vue-loader": "^16.8.3",
-    "vue-tsc": "^0.40.13"
+    "typescript": "~5.3.3",
+    "vite": "^5.0.10",
+    "vitest": "^1.1.0",
+    "vue-tsc": "^1.8.27"
   }
 }
