@@ -1,5 +1,6 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
+import type { Config } from "tailwindcss";
+
+export default {
   content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx,mdx}"],
   theme: {
     colors: {
@@ -129,4 +130,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-};
+} satisfies Config;
