@@ -6,18 +6,22 @@ module.exports = {
     "../src/**/*.stories.mdx",
     "../src/**/*.stories.@(js|jsx|ts|tsx)"
   ],
+
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@storybook/addon-interactions"
+    "@storybook/addon-interactions",
   ],
-  "framework": "@storybook/vue3",
-  "core": {
-    "builder": "@storybook/builder-vite"
+
+  "framework": {
+    name: "@storybook/vue3-vite",
+    options: {}
   },
+
   "features": {
     "storyStoreV7": true
   },
+
   async viteFinal(config) {
     return mergeConfig(config, {
       // Reuse aliases
@@ -32,4 +36,8 @@ module.exports = {
       },
     });
   },
+
+  docs: {
+    autodocs: true
+  }
 }
