@@ -1,6 +1,7 @@
+/* eslint-env node */
 module.exports = {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
-}
+};
