@@ -26,15 +26,25 @@ import SupportIcon from "./icons/IconSupport.vue";
     <template #heading>Tooling</template>
 
     This project is served and bundled with
-    <a href="https://vitejs.dev/guide/features.html" target="_blank">Vite</a>.
-    The recommended IDE setup is
-    <a href="https://code.visualstudio.com/" target="_blank">VSCode</a> +
-    <a href="https://github.com/johnsoncodehk/volar" target="_blank">Volar</a>.
-    If you need to test your components and web pages, check out
-    <a href="https://www.cypress.io/" target="_blank">Cypress</a> and
     <a
-      href="https://docs.cypress.io/guides/component-testing/introduction"
+      href="https://vitejs.dev/guide/features.html"
       target="_blank"
+      rel="noopener"
+      >Vite</a
+    >. The recommended IDE setup is
+    <a href="https://code.visualstudio.com/" target="_blank" rel="noopener"
+      >VSCode</a
+    >
+    +
+    <a
+      href="https://github.com/johnsoncodehk/volar"
+      target="_blank"
+      rel="noopener"
+      >Volar</a
+    >. If you need to test your components and web pages, check out
+    <a href="https://www.cypress.io/" target="_blank" rel="noopener">Cypress</a>
+    and
+    <a href="https://on.cypress.io/component" target="_blank" rel="noopener"
       >Cypress Component Testing</a
     >.
 
@@ -50,13 +60,20 @@ import SupportIcon from "./icons/IconSupport.vue";
     <template #heading>Ecosystem</template>
 
     Get official tools and libraries for your project:
-    <a target="_blank" href="https://pinia.vuejs.org/">Pinia</a>,
-    <a target="_blank" href="https://router.vuejs.org/">Vue Router</a>,
-    <a target="_blank" href="https://test-utils.vuejs.org/">Vue Test Utils</a>,
-    and
-    <a target="_blank" href="https://github.com/vuejs/devtools">Vue Dev Tools</a
+    <a href="https://pinia.vuejs.org/" target="_blank" rel="noopener">Pinia</a>,
+    <a href="https://router.vuejs.org/" target="_blank" rel="noopener"
+      >Vue Router</a
+    >,
+    <a href="https://test-utils.vuejs.org/" target="_blank" rel="noopener"
+      >Vue Test Utils</a
+    >, and
+    <a href="https://github.com/vuejs/devtools" target="_blank" rel="noopener"
+      >Vue Dev Tools</a
     >. If you need more resources, we suggest paying
-    <a target="_blank" href="https://github.com/vuejs/awesome-vue"
+    <a
+      href="https://github.com/vuejs/awesome-vue"
+      target="_blank"
+      rel="noopener"
       >Awesome Vue</a
     >
     a visit.
@@ -69,14 +86,21 @@ import SupportIcon from "./icons/IconSupport.vue";
     <template #heading>Community</template>
 
     Got stuck? Ask your question on
-    <a target="_blank" href="https://chat.vuejs.org">Vue Land</a>, our official
-    Discord server, or
-    <a target="_blank" href="https://stackoverflow.com/questions/tagged/vue.js"
+    <a href="https://chat.vuejs.org" target="_blank" rel="noopener">Vue Land</a
+    >, our official Discord server, or
+    <a
+      href="https://stackoverflow.com/questions/tagged/vue.js"
+      target="_blank"
+      rel="noopener"
       >StackOverflow</a
     >. You should also subscribe to
-    <a target="_blank" href="https://news.vuejs.org">our mailing list</a> and
-    follow the official
-    <a target="_blank" href="https://twitter.com/vuejs">@vuejs</a>
+    <a href="https://news.vuejs.org" target="_blank" rel="noopener"
+      >our mailing list</a
+    >
+    and follow the official
+    <a href="https://twitter.com/vuejs" target="_blank" rel="noopener"
+      >@vuejs</a
+    >
     twitter account for latest news in the Vue world.
   </WelcomeItem>
 
@@ -88,6 +112,8 @@ import SupportIcon from "./icons/IconSupport.vue";
 
     As an independent project, Vue relies on community backing for its
     sustainability. You can help us by
-    <a target="_blank" href="https://vuejs.org/sponsor/">becoming a sponsor</a>.
+    <a href="https://vuejs.org/sponsor/" target="_blank" rel="noopener"
+      >becoming a sponsor</a
+    >.
   </WelcomeItem>
 </template>
