@@ -1,6 +1,4 @@
-/// <reference types="vitest" />
-
-import { fileURLToPath, URL } from "url";
+import { fileURLToPath, URL } from "node:url";
 
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
@@ -10,7 +8,7 @@ export default defineConfig({
   define: {
     __VUE_OPTIONS_API__: false,
   },
-  plugins: [vue({ reactivityTransform: true })],
+  plugins: [vue()],
   resolve: {
     alias: {
       "@": fileURLToPath(new URL("./src", import.meta.url)),
